+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# vNext
+
+- Added better handling for literal types to `ReflectedClass`.
+    * You can now expect `isClass(Boolean)` to be true for types `true` and `false`, `isClass(Object)` to be true for 
+      `null`, `isClass(Number)` to be true for numeric literals and `isClass(String)` to be true for string literals.
+    * Added `isLiteral(value)` to check for a literal value
+
 # v0.0.18
 
 - Added support for type literal types, ie `foo(bar : false, baz : null, foobar : 123)`